mcp-servers/postgres/requirements.txt
mcp[cli]>=1.6.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
